rename directory and files for easier import
