+++ src/components/announcementsList/index.tsx
