+++ src/containers/home/index.tsx
@@ -4,11 +4,11 @@ import styled from 'styled-components';
 import { ORANGE } from '../../utils/colors';
 import { LinkButton } from '../../components/LinkButton';
 import EventCard from '../../components/EventCard';
-import { AnnouncementsDataProps } from '../announcements/Announcements';
+import { AnnouncementsDataProps } from '../announcements';
 import { getAnnouncements } from '../announcements/ducks/thunks';
 import { connect, useDispatch } from 'react-redux';
 import { asyncRequestIsComplete, asyncRequestIsFailed, asyncRequestIsLoading } from '../../utils/asyncRequest';
-import AnnouncementsList from '../../components/announcements-list/AnnouncementsList';
+import AnnouncementsList from '../../components/announcementsList';
 import { C4CState } from '../../store';
 import { Card, Col, Row, Typography } from 'antd';
 const { Text, Paragraph } = Typography;
